Initial support for Newsstand.
This commit enables Newsstand by default: to disable, delete the declaration of BAKER_NEWSSTAND (to remove Newsstand support completely, the Info.plist file should be edited as well).
This also hits a local server to fetch the manifest with the list of issues.
Downloading and reading are not yet supported.