Info/About View for Shelf
I added a modal view that will be presented via an information icon on the top right hand side of the shelf.  This pull request will probably rely on the prior pull request that had the modified Constants.h file.

There are 2 XIB files, one formatted for the iPad and one formatted for the iPhone.  The developer can freely design an about page by using standard Interface Builder controls to create a static about screen that will be presented modally from the button on the top right of the shelf.  There is a button at the top right hand side of the info view that will close the view.  It can be styled (background color and fore color text) via two entries in the Constants.h file.