Added "connecting" before "downloading" issue.
In a Newsstand-enabled shelf, tapping the "download" button will show a "connecting…" message before the "downloading…" message.
This will hopefully make it clear that - when a network connection is missing - download has not started yet, and that it will be started automatically by Newsstand as soon as a connection is available.

Closes #593