Updating Deployment Target
As of February 2014 all apps submitted to the App Store need to be iOS7 compliant.  Changing default deployment target to iOS7.  I believe that developers can still target iOS 6 if required.  However we are changing the default as we believe that it is in line with what Apple is pushing and seeing that iOS7 already has a huge adoption.