Notify Server Backend with Device Type
This addresses issue #14.  Adds the ability to append :devicetype parameter to the API endpoints in Constants.h

Will notify the backend server (as long as it accepts/expects it) of the type of device that is making the call.  Sends either "tablet" or "phone".  Tablet is any type of iPad and Phone is any type of iPhone/iPod Touch