Fixes #403 - Single tap instead of double tap for TOC
Added the baker-toc://show, baker-toc://hide and baker-toc://toggle special urls so that you can show, hide and toggle the navigation bars from within the html.

This allows the user to use plain javascript to accomplish the single tap to show the navigation bars instead of double tap.

We don't do this in the Objective-C part as there is no reliable way to detect a single tap in a web view. It's possible when you are using hyperlinks, but they can't distinguish between a click on a JavaScript element or a weblink.